klighd.view: Added icon for secondary diagram view
The forked views are using a modified icon than the default view. This
way the secondary is easily distinguishable from the primary.
At the moment the icon is not properly adjusted if a secondary view is
in the background at eclipse startup because the view is not initialized
unit it is in the foreground.

Signed-off-by: Nis Wechselberg <nbw@informatik.uni-kiel.de>