klighd: Fixed actions on edges
Actions attached to edges were not triggered. This was caused by the
KlighdSelectionEventHandler that was altering the node stack while
collecting all the selected edges. For this reason the action didn't
receive the proper KGraphElement afterwards.
The fix introduces a new reset() method in piccolo's PPickPath that can
be used to reset the node stack to the original state. This method is
then called after the SelectionEventHandler finished collecting its
data.

Signed-off-by: Nis Wechselberg <nbw@informatik.uni-kiel.de>