klighd: Refactored LightDiagramServices public interface
In LightDiagramServices there were lots of different layoutDiagram()
methodse, all with their own set of configured parameters. All of these
methods were consolidated. The layoutDiagram() (and updateDiagram())
method now takes a LightDiagramLayoutConfig object as parameter which
can be configured according to the various needs of the user. The new
config object is built using fluent interface methods.
The existing layoutDiagram() and updateDiagram() methods have been
deprecated.
The animation time properties need to be passed to elk in the
globalSettings. This hasn't happened since the migration.
Fixes KIPRA-1819.