KlighD: Added handling of animation related options from root node
When using KlighD there currently is no real way to configure the
animation related properties. This used to be done by global
configurations which have been removed in the new elk framework.
The new approach is to define the animation configuration either in the
call to layoutDiagram or, if this is not possible, in the root node.

Signed-off-by: Nis Wechselberg <nbw@informatik.uni-kiel.de>