klighd: Fixed problems with model updating and source model tracking. (KIPRA-1825)
The SimpleUpdateStrategy formerly copied shape layouts over from an upated view model to the current view model. With the new graph, that did not happen anymore, which caused problems for the view model's root node, whose properties never got updated.

The SourceModelTrackingAdapter meanwhile thought it would be helpful to cast something before it knew whether or not the cast would, in fact, succeed. Thank you, SourceModelTrackingAdapter.

Signed-off-by: Christoph Daniel Schulze <cds@informatik.uni-kiel.de>