klighd: Implemented property for clipped node handling
Ports of clipped nodes used to be shown until recently.
Now the KlighdProperty SHOW_CLIPPED_PORTS allows for configuration of
this behavior.

If ports of the clipped node are not shown, the zoom behavior is also
adjusted to show only the children of the clipped node, and not zoom to
the (hidden) ports of the node

Signed-off-by: Nis Wechselberg <nbw@informatik.uni-kiel.de>