constraints: Fix for Reeval, Changes reval, adds setCoordDepthFirst
Bugfix: Reeval uses the parent of the target now instead of the viewModel's root. in order to get the nodes of the same hierachical level. EmptyLayerReevaluation should become an optional reeval now that the system is able to handle offsets in layer constraints. It is commented out for now. setCoordinatesDepthFirst is created. It will be used to fix a layer assignment bug for hierarchical graphs.