constraints: Deletes not needed code, Fixes bug in shifting reeval
Deletes not needed code in getNodesOfLayer(), basically reverting it to an older state. Deletes the getLayering() stub that is not needed. Creates and implements a getAdjacentNodes() method.
Fixes a stackoverflow that was caused by the shift reeval.
Fixes that the shift reeval didn't work with layer constraints > count of layers.